Moved separate files into generla managers file to follow python design
